+++ R/grouped_ggbarstats.R
@@ -11,7 +11,7 @@
 #' @import ggplot2
 #'
 #' @importFrom dplyr select
-#' @importFrom rlang !! enquo quo_name ensym
+#' @importFrom rlang enquo quo_name ensym
 #' @importFrom purrr map
 #'
 #' @seealso \code{\link{ggbarstats}}, \code{\link{ggpiestats}},